# `import --replace-existing` tweaks & optimisations