## Description

When importing features, if there is also a schema change, we compare new blobs against corresponding existing blobs. This is so we can avoid writing subtly-different-but equivalent blobs to the repo unnecessarily.

This is a slow process since we need to fetch all the blobs to do that, so until now we were only running the comparison when a schema change occurred in the commit.

However, during a discussion with @olsen232 we realised that a later non-schema-change commit also needs to run the feature blob comparison, since any features imported with the new schema will vary from the previously-imported feature blob with the old schema.

This PR makes that adjustment (run the feature blob comparison during any import in a repo which has previously had a schema change).

Of necessity I looked into performance to try and mitigate the impact this would have. I found a couple of things which are helpful:

* d1f5dbd9f7f46fc666dfccde6423b69047ecff71 sped up the feature comparison by ~30-40% 👍 
* Increasing the libgit2 object cache size to allow for all sno's tree objects improved the performance by a _further_ ~80-90%.

For reference in a repo with 1M features on my laptop, I achieved the following results for the first 200K features imported in the followup revision:

* no feature comparison, on master: `200,000 features... @11.7s`
* with feature comparison before d1f5dbd9f7f46fc666dfccde6423b69047ecff71: `200,000 features... @68.5s`
* with feature comparison after d1f5dbd9f7f46fc666dfccde6423b69047ecff71: `200,000 features... @47.8s`
* with feature comparison after this PR: `200,000 features... @16.5s`

This also speeds up `sno show` on a large (1M rows imported) commit by 3.5x (4m38 --> 1m45), I guess due to the same issue of trees being looked up repeatedly.

## Related links:

https://github.com/libgit2/libgit2/blob/508361401fbb5d87118045eaeae3356a729131aa/include/git2/common.h#L266

## Checklist:

- [x] Have you reviewed your own change?
- [ ] Have you included test(s)?
- [ ] Have you updated the [changelog](https://github.com/koordinates/sno/blob/master/CHANGELOG.md)?